refactor(connector-tester): extract StructuralMutator from MutationEngine
